+++ packages/storefront/storefront.config.mjs
@@ -1,5 +1,5 @@
 import { fileURLToPath } from 'url';
-import { readFileSync } from 'fs';
+import fs from 'fs';
 import { resolve as resolvePath } from 'path';
 import config from '@cloudcommerce/config';
 
@@ -8,8 +8,6 @@ export default () => {
     STOREFRONT_BASE_DIR,
     VITE_ECOM_STORE_ID,
   } = import.meta.env || process.env;
-
-  const { storeId, lang } = config.get();
 
   let baseDir;
   if (STOREFRONT_BASE_DIR) {
@@ -23,9 +21,24 @@ export default () => {
     config.set({ storeId: Number(VITE_ECOM_STORE_ID) });
   }
 
-  const settings = JSON.parse(
-    readFileSync(resolvePath(dirContent, 'settings.json'), 'utf8'),
-  );
+  const {
+    storeId,
+    lang,
+    countryCode,
+    currency,
+    currencySymbol,
+  } = config.get();
+
+  const cms = (filename) => {
+    if (filename.endsWith('/')) {
+      const dirColl = resolvePath(dirContent, filename);
+      return fs.readdirSync(dirColl).map((_filename) => _filename.replace('.json', ''));
+    }
+    const filepath = resolvePath(dirContent, `${filename}.json`);
+    return JSON.parse(fs.readFileSync(filepath, 'utf8'));
+  };
+
+  const settings = cms('settings');
   const { domain } = settings;
   const primaryColor = settings.primary_color || '#20c997';
   const secondaryColor = settings.secondary_color || '#343a40';
@@ -33,9 +46,13 @@ export default () => {
   return {
     storeId,
     lang,
+    countryCode,
+    currency,
+    currencySymbol,
     domain,
     primaryColor,
     secondaryColor,
     settings,
+    cms,
   };
 };