+++ scripts/version-and-release.sh
@@ -5,4 +5,3 @@ pnpm test
 pnpm run -r prerelease
 npx standard-version --commit-all
 npx zx scripts/release.mjs --publish
-pnpm run -r postrelease